// frontend/src/api/generated/models/routingStagePayload.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */
import type { RoutingStagePayloadTransitionRule } from './routingStagePayloadTransitionRule.ts';

export interface RoutingStagePayload {
  sequence: number;
  stage_name: string;
  transition_rule?: RoutingStagePayloadTransitionRule;
  requires_qc?: boolean;
}