// frontend/src/api/generated/models/routingStageResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */
import type { RoutingStageResponseTransitionRule } from './routingStageResponseTransitionRule.ts';

export interface RoutingStageResponse {
  id: string;
  sequence: number;
  stage_name: string;
  transition_rule: RoutingStageResponseTransitionRule;
  requires_qc: boolean;
}